Changed shiftfs module mount sequence to allow mounting it from outside the source user-ns.
This change allows shiftfs mounts to be done in a single step that
specifies both of the user namespaces involved in the uid/gid
shifting (as opposed to the prior version of shiftfs which requires
two steps, marking and mounting).

This change allows us to fix the problems with mounting shiftfs on
container volumes for which the container's init process does not have
access. See the shiftfs README.md file for details.