Bug fix and minor improvements to sysfs /proc/sys inter-container access test.
The test had a bug in a loop in which it was supposed to iterate
through all sys containers are execute a command, yet was only
executing the command on one sys container.

Also, improved the test by using "docker cp" to copy a worker script'
into the sys container instead of bind-mounting it.