Add the linux headers package as a dependency in the Sysbox Debian/Ubuntu installer.
In order to run K8s inside sys containers, Sysbox must mount the
host's kernel headers directory onto the system container (because K8s
peeks into the kernel's configuration file present in this directory).

This necessarily means that the kernel-headers must be present in the
host machine. Ubuntu hosts include the kernel headers by
default. Debian hosts need to explicitly install the linux-headers
package.

This commit makes the Sysbox installer for Debian/Ubuntu check that
the Linux kernel headers are installed on the host. If they aren't,
Sysbox installation will fail.

In addition, this change modifies the Dockerfile for the installer
test container to include the kernel headers package (as otherwise
the Sysbox installation inside the test container fails).