Fix for Helm-v2 integration-test instability
Problem prevented Helm-v2 test to pass, which also ended up causing the failure of Helm-v3 testcase.

Ultimate root-cause was an unexpected POD being present during the execution of Helm-v2. This POD was/is instantiated by the previous testcase (node-down), and it was/is unable to eliminate the POD in question for quite a while.

Prior to this PR, Helm-v2/v3 testcases operated under the assumption that no other (user) PODs were present in the cluster, which is not a valid assumption based on the above.

The fix is to narrow down the logic in Helm-v2 testcase to match the specific PODs expected by this testcase.