Add test to verify that multiple sys containers can be started in parallel.
This test verifies that multiple sys containers can be started in parallel.
The test uses sys containers that come preloaded with heavy inner container
images to put strain on Sysbox. It also launches twice the number of sys
containers as there are CPUs on the host, to add further strain.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>