Refactor Makefile's kernel-path-identification logic to allow various distros [skip ci]
These changes are part of our effort to expand the number of Linux distros supported by Sysbox. Here we are extending the logic responsible for the identification of the kernel-header's path so that we allow Sysbox compilation/testing in distros other than Ubuntu.

Note: I'm skiping CI here as i'll be generating a new PR right after to update sysbox-runc submodules.

Signed-off-by: Rodny Molina <rmolina@nestybox.com>