Increase the size of the kernel.keys.maxbytes sysctl in sysbox tests.
The sysbox test framework increases the kernel.keys.maxkeys to 20K (in
order to accommodate a large number of sys containers), yet the
kernel.keys.maxbytes was not increased accordingly.

As a result of this discrepancy, tests that create many sys containers
in parallel were failing (e.g., the "tests/perf/basic.bats" test failed
executed on machines with large number of CPU cores).

This change fixes this.

Signed-off-by: ctalledo <ctalledo@nestybox.com>