Change Sysbox tests to use container images in Nestybox's GitHub Container Registry.
As of late 2020, Dockerhub imposes rate limiting on pulls for anonymous
users. This is causing problems for the Sysbox CI, since the CI tests run inside
a privileged container running as an anonymous Docker user. The problem is that
after the Dockerhub rate limis are exceeded, the CI tests start failing with
errors such as:

docker: Error response from daemon: toomanyrequests: You have reached your pull rate limit.

To work-around this, this commit modifies the tests to use Nestybox's GitHub
Container Registry (ghcr.io), which has the same images as the Dockerhub repo.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>