Revert part of the changes made to testContainerInit [skip ci]
The previous changes being reverted attempted to avoid the execution of testContainerInit's docker-configuration logic in the scenario where the Sysbox installer was utilized. My purpose then was to
exclusively rely on the installer's config scripts to verify its proper operation.

However, by doing that we were missing to configure docker inside the sysbox-test container as requested by the sysbox Makefile, as the Sysbox installer is not aware of the userns-vs-shiftfs env-vars
passed into the sysbox-test container during initialization. In consequence, dockerd was always operating in userns mode.

The side-effect of reverting these changes is that now we are not fully testing the installer's configuration logic, as most of the docker-config elements are already in place by the time the installer runs.

Signed-off-by: Rodny Molina <rmolina@nestybox.com>