Add a helper script to find the kernel stack trace for a process tree. [skip ci]
This is helpful when debugging system container processes, in particular to see
if any of them are stuck / hanging in the kernel waiting for an operation
trapped by Sysbox.

Sample usage:

root@sysbox-test:~/nestybox/sysbox# ./scr/pid_bt some_pid

 --- PID 4472 TID 4472 ---
[<0>] futex_wait_queue_me+0xbb/0x120
[<0>] futex_wait+0x105/0x260
[<0>] do_futex+0x146/0x4c0
[<0>] __x64_sys_futex+0x13f/0x170
[<0>] do_syscall_64+0x57/0x190
[<0>] entry_SYSCALL_64_after_hwframe+0x44/0xa9

 --- PID 4472 TID 4474 ---
[<0>] futex_wait_queue_me+0xbb/0x120
[<0>] futex_wait+0x105/0x260
[<0>] do_futex+0x146/0x4c0
[<0>] __x64_sys_futex+0x13f/0x170
[<0>] do_syscall_64+0x57/0x190
[<0>] entry_SYSCALL_64_after_hwframe+0x44/0xa9

...

If any of those shows "fuse" or any of the syscalls trapped by
sysbox (e.g., mount, etc.), then it means the process is likely
stuck waiting for Sysbox.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>