Update Golang (1.13.3 -> 1.16.4) and Grpc-protobuf (3.6.1 -> 3.15.8) in test-server dockerfiles
Update Golang to take advantage of the multiple enhancements made over the last couple of years, particularly in the performance and concurrency areas. The chosen release is the penultimate 1.16 stable release.

The GRPC upgrade is a side-effect due to some incompatibility issues detected (with the old Golang release) during the protobuf compiler installation/building process. Either way, it was about time that we upgrade grpc/protobuf modules too. The chosen release is the stable one recommended/utilized by the official Grpc installation [guide](https://grpc.io/docs/protoc-installation/) -- the latest release is 3.17.3, so we are being relatively conservative here.

https://golang.org/doc/go1.14
https://golang.org/doc/go1.15
https://golang.org/doc/go1.16

Signed-off-by: Rodny Molina <rmolina@nestybox.com>