Fix failure of cgroups v1 test in Debian Buster machines.
The cgroups memory.failcnt counter behaves different in Debian Buster compared
to other distros. This is not a big deal, but was causing a cgroups v1 test to
fail.  This commit modifies that test to prevent the failure.

Tested on both a Debian Buster host as well as on a Ubuntu Focal host.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>