Update sysbox-deploy k8s manifests, per latest changes in the sysbox-deploy-k8s daemonset.
The latest version of the sysbox-deploy-k8s daemonset also performs the CRI-O
installation, thus voiding the need for a separate crio-installation daemonset.
This simplifies Sysbox installation on K8s clusters and makes it twice as fast.

This commit has the required changes to the sysbox-deploy k8s manifests for this.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>