Modify KinD-in-Sysbox tests to remove reliance on kindind image.
The nestybox/kindind image includes the K8s.io KinD tool inside, and also comes
with the inner nestybox/kindestnode image preloaded.  This is useful as a
reference example of how to run KinD inside Sysbox.

However, most users will probably not use the nestybox/kindind image to deploy a
container with KinD inside; instead they will create a sysbox container with an
image of their own, install KinD and kubectl, and deploy the KinD cluster
inside.

Therefore, this commit modifies the kindind tests to use the approach
most likely to be used by users.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>