Docs: Increase the 'sleep' interval in sysbox-k8s' uninstallation process [skip ci]
The previous increment (0 -> 30 secs) fixed the main issue that prevented Kubelet from switching from CRI-O back to its original CNI (i.e. dockershim or containerd). In this case we are not addressing a functional issue as there's no service impact per-se, but just a cosmetic one. In certain scenarios, the "sysbox-runtime=running" label may not be eliminated during the uninstallation cycle, which is just a side-effect of the 'kubelet-unconfig-helper' running slower than usual. As a result, the "sysbox-runtime" label will continue to be associated with a node that isn't running sysbox any longer.

The fix is to simply extend the 'sleep' interval another 30 secs.

Signed-off-by: Rodny Molina <rmolina@nestybox.com>