First commit of sysbox-docker-cp tool.
This tool is a wrapper around "docker cp" meant to be used in containers
launched with Docker + Sysbox.

This tool is needed because in some cases, "docker cp" does not work well as it
does not adjust the uid:gid of the file copied based on the container's
user-namespace uid/gid mappings.

Specifically, "docker cp" does not work well in hosts that don't support
shiftfs. The reason is that in such hosts, Sysbox is forced to shift the
container's rootfs uid:gids to match the container's assigned user-ns uid/gid
mappings using chown. As a result, it follows that "docker cp" should also do
the chown, but since Docker is not aware that Sysbox has enabled user-ns on the
container, "docker cp" won't do the chown (unless Docker is configured in
userns-remap mode).

The solution we came up is to create sysbox-docker-cp to perform the "docker cp"
and the chown of the copied file correctly.

Note that the idmapped mounts kernel feature does not help currently, since
idmapped mounts are not supported over rootfs on overlayfs (which is the default
for Docker). Once idmapped mounts are supported on overlayfs, this tool won't
be needed anymore: "docker cp" need not chown anything, since the container's
idampped-mount of the rootfs will cause the kernel to do the uid:gid shifting
(just as with shiftfs).

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>