Deprecate *-shiftuid test targets (no longer needed).
This commit deprecates the *-shiftuid test targets in the Makefile.

The reason is that with the addition of idmapped mount support to Sysbox, it
was apparent that these Make targets did not scale to the testing needs.

A better way to do this is to launch a test target such as "make test-sysbox",
and let the test scripts configure Sysbox (and Docker) in the desired modes
of operation (e.g., with idmapped mount, with shiftfs, with Docker userns-remap,
and combinations of these).

With this change, a single invocation of "make test-sysbox" will run the
Sysbox test suite and verify all the operating modes described above
(see the test scripts at tests/scr/testSysbox for an example of how
Sysbox and Docker are reconfigured during the test to cover each
of these modes).

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>