Fix testing issue preventing 'sysfs' testcases from passing
Missed updating the baseline images that we are using for sysfs testing (i.e. busybox and debian), and this must be done attending to the 'arch' being used in the build-server as otherwise tests won't run (obviously).

I spent a while trying to build 'dynamic' rootfs for each of these images (to avoid attaching another 40MB file to our repo), but I couldn't make 'docker export' work as part of 'tests/helpers/setup.bash' -- looks like a 'bats' limitation of some sort.

Signed-off-by: Rodny Molina <rmolina@nestybox.com>