Shorten Sysbox CI tests.
The test suite it's currently taking several hours per run, as it's testing
Sysbox in all combinations of ID-mapped mounts + shiftfs.

In this commit, we test only in one combination: either ID-mapped mounts +
shiftfs (if the test host supports both), or ID-mapped mounts only, or shiftfs
only.

Since we have CI runners that support either ID-mapped mounts or shiftfs, we
will actually run Sysbox in all combinations, so the impact on test coverage is
rather small.

Also, this commit reduces the number of test that we run in the CI suite.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>