Ensure no build artifacts are left with host root ownership.
Prior to this change, some build artifacts had host root:root
ownership, due to them being build inside the sysbox test
container, which is a privileged container with root as the
default user.

This change fixes this chowing any build artifacts to
be owned by the uid of the user running the sysbox
test container.

As a result of this change, "make clean" now works
without requiring "sudo".

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>