Adjust tests for ID-mapping of rootfs and implicit mounts.
Now that Sysbox supports ID-mapping the container's rootfs (in Linux kernel
5.19+), we need to adjust tests and add new ones accordingly.

Same applies for ID-mapping of Sysbox volumes that are implicitly mounted into
the container. In the past those could not be ID-mapped because programs running
inside the container would typically mount overlayfs in those dirs. However in
Linux kernel 5.19+ they can be ID-mapped, so Sysbox takes advantage of this and
the tests need to be modified accordingly.

Signed-off-by: Cesar Talledo <cesar.talledo@docker.com>