Make tests smarter when checking container capabilities.
Rather than assuming that a root process in a sys container has capabilities of
000001ffffffffff, get the max set of capbilities for a root process in the test
container and compare against those.

This way, if Linux adds new capabilities, or if the capabilities of a root
process in the test container change due to a change in the host's Docker, the
test will adjust appropriately.

Signed-off-by: Cesar Talledo <cesar.talledo@docker.com>