Fix race-condition in sysKernel integration-testcase
The "/sys/kernel/mm/ksm file ops" testcase is failing every now and then when comparing nodes inside and outside the sys-container. This is expected since some of these nodes are tracking system resources (e.g., "/sys/kernel/mm/ksm/pages_unshared").

The fix is to iterate a few times through the affected node to ensure that failure is not a triggered by a transient issue.

Signed-off-by: Rodny Molina <rodny.molina@docker.com>